+++ framework_src/native_library/FreSwiftExampleANE/FreSwift/FlashRuntimeExtensions.swift
@@ -53,86 +53,6 @@ public extension FREObject {
         return ""
     }
     
-    /// getProp: returns the Property of a FREObject.
-    ///
-    /// ```swift
-    /// let myName = argv[0].getProp("name")
-    /// ```
-    /// - parameter name: name of the property to return
-    /// - throws: Can throw a `FreError` on fail
-    /// - returns: FREObject?
-    @available(*, unavailable, message: "use accessor or FreObjectSwift wrapper instead")
-    func getProp(name: String) throws -> FREObject? {
-        if let ret = FreSwiftHelper.getProperty(rawValue: self, name: name) {
-            return ret
-        }
-        return nil
-    }
-
-    /// setProp: sets the Property of a FREObject.
-    /// - parameter name: name of the property to set
-    /// - parameter value: value to set to
-    /// - throws: Can throw a `FreError` on fail
-    /// - returns: Void
-    @available(*, unavailable, message: "use accessor or FreObjectSwift wrapper instead")
-    func setProp(name: String, value: Any?) throws {
-        if value is FREObject {
-            FreSwiftHelper.setProperty(rawValue: self, name: name, prop: value as? FREObject)
-        } else if value is FREArray {
-            if let v = value as? FREArray {
-              FreSwiftHelper.setProperty(rawValue: self, name: name, prop: v.rawValue)
-            }
-        } else if value is FreObjectSwift {
-            if let v = value as? FreObjectSwift {
-                FreSwiftHelper.setProperty(rawValue: self, name: name, prop: v.rawValue)
-            }
-        } else if value is String {
-            if let v = value as? String {
-                FreSwiftHelper.setProperty(rawValue: self, name: name, prop: v.toFREObject())
-            }
-        } else if value is Int {
-            if let v = value as? Int {
-                FreSwiftHelper.setProperty(rawValue: self, name: name, prop: v.toFREObject())
-            }
-        } else if value is Int32 {
-            if let v = value as? Int32 {
-                FreSwiftHelper.setProperty(rawValue: self, name: name, prop: (Int(v)).toFREObject())
-            }
-        } else if value is UInt {
-            if let v = value as? UInt {
-                FreSwiftHelper.setProperty(rawValue: self, name: name, prop: v.toFREObject())
-            }
-        } else if value is UInt32 {
-            if let v = value as? UInt32 {
-                FreSwiftHelper.setProperty(rawValue: self, name: name, prop: (UInt(v)).toFREObject())
-            }
-        } else if value is Double {
-            if let v = value as? Double {
-                FreSwiftHelper.setProperty(rawValue: self, name: name, prop: v.toFREObject())
-            }
-        } else if value is CGFloat {
-            if let v = value as? CGFloat {
-                FreSwiftHelper.setProperty(rawValue: self, name: name, prop: v.toFREObject())
-            }
-        } else if value is Bool {
-            if let v = value as? Bool {
-                FreSwiftHelper.setProperty(rawValue: self, name: name, prop: v.toFREObject())
-            }
-        } else if value is Date {
-            if let v = value as? Date {
-                FreSwiftHelper.setProperty(rawValue: self, name: name, prop: v.toFREObject())
-            }
-        } else if value is CGRect {
-            if let v = value as? CGRect {
-                FreSwiftHelper.setProperty(rawValue: self, name: name, prop: v.toFREObject())
-            }
-        } else if value is CGPoint {
-            if let v = value as? CGPoint {
-                FreSwiftHelper.setProperty(rawValue: self, name: name, prop: v.toFREObject())
-            }
-        }
-    }
-    
     /// init: Creates a new FREObject.
     ///
     /// ```swift
@@ -215,7 +135,7 @@ public extension FREObject {
     /// value: returns the Swift value of a FREObject.
     ///
     /// - returns: Any?
-    public var value: Any? {
+    var value: Any? {
         return FreObjectSwift(self).value
     }
 }
@@ -229,9 +149,7 @@ public extension NSNumber {
     /// - parameter freObject: FREObject which is of AS3 type Number
     /// - returns: NSNumber?
     convenience init?(_ freObject: FREObject?) {
-        guard let rv = freObject else {
-            return nil
-        }
+        guard let rv = freObject else { return nil }
         if let d = FreSwiftHelper.getAsDouble(rv) {
             self.init(value: d)
         } else {
@@ -258,9 +176,7 @@ public extension Double {
     /// - parameter freObject: FREObject which is of AS3 type Number
     /// - returns: Double?
     init?(_ freObject: FREObject?) {
-        guard let rv = freObject else {
-            return nil
-        }
+        guard let rv = freObject else { return nil }
         if let d = FreSwiftHelper.getAsDouble(rv) {
             self.init(d)
         } else {
@@ -288,9 +204,7 @@ public extension Float {
     /// - parameter freObject: FREObject which is of AS3 type Number
     /// - returns: Float?
     init?(_ freObject: FREObject?) {
-        guard let rv = freObject else {
-            return nil
-        }
+        guard let rv = freObject else { return nil }
         if let d = FreSwiftHelper.getAsDouble(rv) {
             self.init(Float(d))
         } else {
@@ -319,9 +233,7 @@ public extension CGFloat {
     /// - parameter freObject: FREObject which is of AS3 type Number
     /// - returns: CGFloat?
     init?(_ freObject: FREObject?) {
-        guard let rv = freObject else {
-            return nil
-        }
+        guard let rv = freObject else { return nil }
         if let d =  FreSwiftHelper.getAsDouble(rv) {
             self.init(d)
         } else {
@@ -350,9 +262,7 @@ public extension Bool {
     /// - parameter freObject: FREObject which is of AS3 type Boolean
     /// - returns: Bool?
     init?(_ freObject: FREObject?) {
-        guard let rv = freObject else {
-            return nil
-        }
+        guard let rv = freObject else { return nil }
         if let b = FreSwiftHelper.getAsBool(rv) {
             self.init(b)
         } else {
@@ -379,9 +289,7 @@ public extension Date {
     /// - parameter freObject: FREObject which is of AS3 type Date.
     /// - returns: Date?
     init?(_ freObject: FREObject?) {
-        guard let rv = freObject else {
-            return nil
-        }
+        guard let rv = freObject else { return nil }
         if let d = FreSwiftHelper.getAsDate(rv) {
             self.init(timeIntervalSince1970: d.timeIntervalSince1970)
         } else {
@@ -408,9 +316,7 @@ public extension Int {
     /// - parameter freObject: FREObject which is of AS3 type int
     /// - returns: Int?
     init?(_ freObject: FREObject?) {
-        guard let rv = freObject else {
-            return nil
-        }
+        guard let rv = freObject else { return nil }
         if let i = FreSwiftHelper.getAsInt(rv) {
             self.init(i)
         } else {
@@ -438,9 +344,7 @@ public extension UInt {
     /// - parameter freObject: FREObject which is of AS3 type uint
     /// - returns: UInt?
     init?(_ freObject: FREObject?) {
-        guard let rv = freObject else {
-            return nil
-        }
+        guard let rv = freObject else { return nil }
         if let i = FreSwiftHelper.getAsInt(rv) {
             self.init(i)
         } else {
@@ -469,9 +373,7 @@ public extension String {
     /// - parameter freObject: FREObject which is of AS3 type String
     /// - returns: String?
     init?(_ freObject: FREObject?) {
-        guard let rv = freObject, FreSwiftHelper.getType(rv) == FreObjectTypeSwift.string else {
-            return nil
-        }
+        guard let rv = freObject, FreSwiftHelper.getType(rv) == FreObjectTypeSwift.string else { return nil }
         if let s = FreSwiftHelper.getAsString(rv) {
             self.init(s)
         } else {
@@ -501,9 +403,7 @@ public extension String {
         /// - parameter hasAlpha: Whether the uint is in 32bit ARGB hex format ef 0xFF00FF00
         /// - returns: NSColor?
         convenience init?(_ freObject: FREObject?, hasAlpha: Bool = true) {
-            guard let rv = freObject else {
-                return nil
-            }
+            guard let rv = freObject else { return nil }
             guard let fli = CGFloat(rv) else { return nil }
             let rgb = Int(fli)
             let a = hasAlpha ? (rgb >> 24) & 0xFF : 255
@@ -515,16 +415,6 @@ public extension String {
             let gFl = CGFloat(g) / 255
             let bFl = CGFloat(b) / 255
             self.init(red: rFl, green: gFl, blue: bFl, alpha: aFl)
-        }
-        
-        @available(*, obsoleted: 3.0.0, message: "Removed use init(_ freObject: FREObject?, hasAlpha: Bool = true) instead")
-        convenience init?(freObject: FREObject?, alpha: FREObject?) {
-            self.init()
-        }
-        
-        @available(*, obsoleted: 3.0.0, message: "Removed use init(_ freObject: FREObject?, hasAlpha: Bool = true) instead")
-        convenience init?(freObjectARGB: FREObject?) {
-            self.init()
         }
         
         /// toFREObject: Converts a NSColor into a FREObject of AS3 type uint (ARGB).
@@ -553,9 +443,7 @@ public extension String {
         /// - parameter hasAlpha: Whether the uint is in 32bit ARGB hex format ef 0xFF00FF00
         /// - returns: UIColor?
         convenience init?(_ freObject: FREObject?, hasAlpha: Bool = true) {
-            guard let rv = freObject else {
-                return nil
-            }
+            guard let rv = freObject else { return nil }
             guard let fli = CGFloat(rv) else { return nil }
             let rgb = Int(fli)
             let a = hasAlpha ? (rgb >> 24) & 0xFF : 255
@@ -567,16 +455,6 @@ public extension String {
             let gFl = CGFloat(g) / 255
             let bFl = CGFloat(b) / 255
             self.init(red: rFl, green: gFl, blue: bFl, alpha: aFl)
-        }
-        
-        @available(*, obsoleted: 3.0.0, message: "Removed use init(_ freObject: FREObject?, hasAlpha: Bool = true) instead")
-        convenience init?(freObject: FREObject?, alpha: FREObject?) {
-            self.init()
-        }
-        
-        @available(*, obsoleted: 3.0.0, message: "Removed use init(_ freObject: FREObject?, hasAlpha: Bool = true) instead")
-        convenience init?(freObjectARGB: FREObject?) {
-            self.init()
         }
         
         /// toFREObject: Converts a UIColor into a FREObject of AS3 type uint (ARGB).
@@ -608,9 +486,7 @@ public extension Dictionary where Key == String, Value == Any {
     /// - returns: Dictionary?
     init?(_ freObject: FREObject?) {
         self.init()
-        guard let rv = freObject else {
-            return
-        }
+        guard let rv = freObject else { return }
         if let val: [String: Any] = FreSwiftHelper.getAsDictionary(rv) {
             self = val
         }
@@ -627,9 +503,7 @@ public extension Dictionary where Key == String, Value == AnyObject {
     /// - returns: Dictionary?
     init?(_ freObject: FREObject?) {
         self.init()
-        guard let rv = freObject else {
-            return
-        }
+        guard let rv = freObject else { return }
         if let val: [String: AnyObject] = FreSwiftHelper.getAsDictionary(rv) {
             self = val
         }
@@ -646,9 +520,7 @@ public extension Dictionary where Key == String, Value == NSObject {
     /// - returns: Dictionary?
     init?(_ freObject: FREObject?) {
         self.init()
-        guard let rv = freObject else {
-            return
-        }
+        guard let rv = freObject else { return }
         if let val: [String: NSObject] = FreSwiftHelper.getAsDictionary(rv) {
             self = val
         }